feat: socketio in waitscene
