+++ src/index.js
@@ -30,6 +30,7 @@ import BoardDice from './scenes/BoardDice'
 
 import config from './config/config'
 
+
 class Game extends Phaser.Game {
   constructor() {
     // Add the config file to the game
@@ -50,81 +51,6 @@ class Game extends Phaser.Game {
 
     // Start the game with the mainscene
     this.scene.start('WaitScene')
-  }
-
-  create() {
-    var self = this;
-    this.socket = io();
-    this.otherPlayers = this.physics.add.group();
-    this.socket.on('currentPlayers', function (players) {
-      Object.keys(players).forEach(function (id) {
-        if (players[id].playerId === self.socket.id) {
-          this.addPlayer(self, players[id]);
-        } else {
-          addOtherPlayers(self, players[id]);
-        }
-      });
-    });
-    this.socket.on('newPlayer', function (playerInfo) {
-      addOtherPlayers(self, playerInfo);
-    });
-    this.socket.on('disconnect', function (playerId) {
-      self.otherPlayers.getChildren().forEach(function (otherPlayer) {
-        if (playerId === otherPlayer.playerId) {
-          otherPlayer.destroy();
-        }
-      });
-    });
-
-    this.socket.on('playerMoved', function (playerInfo) {
-      self.otherPlayers.getChildren().forEach(function (otherPlayer) {
-        if (playerInfo.playerId === otherPlayer.playerId) {
-          otherPlayer.setRotation(playerInfo.rotation);
-          otherPlayer.setPosition(playerInfo.x, playerInfo.y);
-        }
-      });
-    });
-
-    // this.blueScoreText = this.add.text(16, 16, '', { fontSize: '32px', fill: '#0000FF' });
-    // this.redScoreText = this.add.text(584, 16, '', { fontSize: '32px', fill: '#FF0000' });
-
-    this.socket.on('scoreUpdate', function (scores) {
-      self.blueScoreText.setText('Blue: ' + scores.blue);
-      self.redScoreText.setText('Red: ' + scores.red);
-    });
-
-    this.socket.on('starLocation', function (starLocation) {
-      if (self.star) self.star.destroy();
-      self.star = self.physics.add.image(starLocation.x, starLocation.y, 'star');
-      self.physics.add.overlap(self.ship, self.star, function () {
-        this.socket.emit('starCollected');
-      }, null, self);
-    });
-
-    this.cursors = this.input.keyboard.createCursorKeys();
-  }
-
-  addPlayer(self, playerInfo) {
-    self.ship = self.physics.add.image(playerInfo.x, playerInfo.y, 'ship').setOrigin(0.5, 0.5).setDisplaySize(53, 40);
-    if (playerInfo.team === 'blue') {
-      self.ship.setTint(0x0000ff);
-    } else {
-      self.ship.setTint(0xff0000);
-    }
-    self.ship.setDrag(100);
-    self.ship.setAngularDrag(100);
-    self.ship.setMaxVelocity(200);
-  }
-
-  addOtherPlayers(self, playerInfo) {
-    const otherPlayer = self.add.sprite(playerInfo.x, playerInfo.y, 'otherPlayer').setOrigin(0.5, 0.5).setDisplaySize(53, 40);
-    if (playerInfo.team === 'blue') {
-      otherPlayer.setTint(0x0000ff);
-    } else {
-      otherPlayer.setTint(0xff0000);
-    }
-    otherPlayer.playerId = playerInfo.playerId;
-    self.otherPlayers.add(otherPlayer);
   }
 
 }
