+++ src/scenes/WaitScene.js
@@ -1,5 +1,7 @@
 import 'phaser'
 
+import io from 'socket.io-client';
+
 export default class WaitScene extends Phaser.Scene {
   constructor() {
     super('WaitScene');
@@ -9,6 +11,89 @@ export default class WaitScene extends Phaser.Scene {
     // << LOAD BACKGROUND AND FOREGROUND SCENES IN PARALLEL HERE >>
     this.scene.launch('WaitBg');
     this.scene.launch('WaitFg');
+
+    console.log("in create")
+    var self = this;
+    this.socket = io('http://localhost:3000');
+    this.otherPlayers = this.physics.add.group();
+
+    this.socket.on('currentPlayers', function (players) {
+      Object.keys(players).forEach(function (id) {
+        if (players[id].playerId === self.socket.id) {
+          self.addPlayer(self, players[id]);
+        } else {
+          self.addOtherPlayers(self, players[id]);
+        }
+      });
+    });
+
+    this.socket.on('connect', function(){
+      console.log("client - connected");
+
+      self.input.keyboard.on('keydown_W', function() {
+        self.socket.emit('testKey')
+      });
+
+
+    });
+    this.socket.on('newPlayer', function (playerInfo) {
+      self.addOtherPlayers(self, playerInfo);
+    });
+    this.socket.on('disconnect', function (playerId) {
+      self.otherPlayers.getChildren().forEach(function (otherPlayer) {
+        if (playerId === otherPlayer.playerId) {
+          otherPlayer.destroy();
+        }
+      });
+    });
+
+    this.socket.on('playerMoved', function (playerInfo) {
+      self.otherPlayers.getChildren().forEach(function (otherPlayer) {
+        if (playerInfo.playerId === otherPlayer.playerId) {
+          otherPlayer.setRotation(playerInfo.rotation);
+          otherPlayer.setPosition(playerInfo.x, playerInfo.y);
+        }
+      });
+    });
+
+
+    this.socket.on('scoreUpdate', function (scores) {
+      self.blueScoreText.setText('Blue: ' + scores.blue);
+      self.redScoreText.setText('Red: ' + scores.red);
+    });
+
+    this.socket.on('starLocation', function (starLocation) {
+      if (self.star) self.star.destroy();
+      self.star = self.physics.add.image(starLocation.x, starLocation.y, 'star');
+      self.physics.add.overlap(self.ship, self.star, function () {
+        this.socket.emit('starCollected');
+      }, null, self);
+    });
+
+    this.cursors = this.input.keyboard.createCursorKeys();
   }
 
+  addPlayer(self, playerInfo) {
+    self.ship = self.physics.add.image(playerInfo.x, playerInfo.y, 'ship').setOrigin(0.5, 0.5).setDisplaySize(53, 40);
+    if (playerInfo.team === 'blue') {
+      self.ship.setTint(0x0000ff);
+    } else {
+      self.ship.setTint(0xff0000);
+    }
+    self.ship.setDrag(100);
+    self.ship.setAngularDrag(100);
+    self.ship.setMaxVelocity(200);
+  }
+
+  addOtherPlayers(self, playerInfo) {
+    const otherPlayer = self.add.sprite(playerInfo.x, playerInfo.y, 'otherPlayer').setOrigin(0.5, 0.5).setDisplaySize(53, 40);
+    if (playerInfo.team === 'blue') {
+      otherPlayer.setTint(0x0000ff);
+    } else {
+      otherPlayer.setTint(0xff0000);
+    }
+    otherPlayer.playerId = playerInfo.playerId;
+    self.otherPlayers.add(otherPlayer);
+  }
 }
+
